Yeet-0  Beet - O
I changed the main menu so that it scaled with screen size, rather than being constant pixel size.  I also resized somethings, and I fixed the issue I was having with the menu.  I think it might have had something to do with the text box for the title being nearly the size of the entire canvas, but I'm not sure.