Camera Adjustments + Build Tests
I worked on level for a bit, I'm also considering moving level 3 to be a challenge level because its so difficult.  I still can't beat it but I know it's possible.  I made adjustments to the ccanvas scalers and the camera script so now it works when building it.  We're almost done dad